Collectd+InfluxDb-RMON Replacement(ALL METRICS) P1
This update introduces two new open source services.

The first is 'collectd' which is a plugin based resource monitoring
engine intended to obsolete the existing rmond process.
The second service is 'influxdb' which is a light weight time series
database intended to store sample data.

Collectd and influxdb are configured through puppet and are started by
manifest.

Collectd runs on all hosts while influxdb only runs on the controllers,.
Colectd periodically monitors platform cpu, memory and filesystem
resources in terms of percentage used value. The value sample for each
monitored resource for each audit cycle is forwarded to the active
controller for storage in the 'influxdb' database.

This update develops several collectd 'support' plugins for managing
thresholds, alarms, and host degrade. Additionally, this update
implements home-grown plugins for platform cpu and memory monitoring
and combines the static and dynamic filesystem resource monitoring in
rmond into a single filesystem plugin based on collectd's stock 'df'
plugin.

Collectd's network plugin is configured to forward samples to the active
controller while influxdb that only runs on the controllers are
configured to listen for collectd samples and then store them in its
database.

Change-Id: I13fd033bb1d14f299dcb97fa80296641c958d0a9
Signed-off-by: Jack Ding <jack.ding@windriver.com>