Mtce: mtcAgent sometimes coredumps on process exit
The mtcAgent process has been seen to segfault and coredump on process exit.

The exit code is interating over a c++ list that can change due to http
interrupt response handling.

The dump code is commented out with a note indicating why and when it could
be re-enabled.

Change-Id: Ie4ef684a65ded533c347ae07fdfa47f332412f7d
Signed-off-by: David Sullivan <david.sullivan@windriver.com>
Story: 2002994
Task: 23008