mtce-storage: implement makefile for installation
mtce-storage doesn't have a Makefile to handle the installation,
and instead, it was directly written in %install inside spec file.
For the sake of multi-os support, it is required to decouple the
installation out of centos spec file. By putting a "install" target
in Makefile, multi-OSs (including CentOS) can take advantage of it.

Story: 2004013
Task: 26993

Change-Id: Id2bda6a2b6afd21aed9897d8d3a5bb211521f639
Signed-off-by: Yong Hu <yong.hu@intel.com>