stx-metal: implement makefile for installation
In these packages (mtce-storage, mtce-compute, mtce-control),
there is no a Makefile to handle the installation, and instead,
the installation steps were directly written in %install inside
original Centos spec file. For the sake of multi-os support, it
is required to decouple the installation out of  spec file.
By making a "install_non_bb" target in Makefile for installatin
steps, multi-OSs (including CentOS) can take advantage of it.

In addition, remove "cgts-" prefix in files in these 3 packages,
to align with the naming convention with other packages.

Story: 2004013
Task: 26991 (for mtce-compute)
Task: 26992 (for mtce-control)
Task: 26993 (for mtce-storage)

Change-Id: Id2bda6a2b6afd21aed9897d8d3a5bb211521f639
Signed-off-by: Yong Hu <yong.hu@intel.com>