get rid of duplicate LICENSE files in 3 packages
There are 2 duplicated LICESE files in mtce-control, mtce-compute,
and mtce-storage. Additionally, LICENSE was not placed in the root
directory of src RPM, so this patch is made as an enhancement or fix.

Test method: make a clean build and check src RPM and binary RPM
to assure there is only one LICENSE in correct place.

Story: 2004186
Task: 27676

Change-Id: Id71a7450e8b45438c5d15976ae8e853b9ba8f4f5
Signed-off-by: Yong Hu <yong.hu@intel.com>