refactor openvswtich
openvswitch-config is created for openvswitch config file
configuration. Add it to the filter the same as openvswitch.

openstack-neutron-openvswitch is filtered from controller node
also. It is a Neutron plugin that implements virtual
networks using Open vSwitch. Should be for compute node only.
Otherwise, due to it dependes on openvswitch, it will causes
openvswitch is installed in controller node also.

Test:
Pass build and multi node deploy test.

Story: 2003768
Task: 27601

Change-Id: Ieaaa1e5db31bca5262cbbcd61ebfc0390be10528
Signed-off-by: slin14 <shuicheng.lin@intel.com>