Remove include bits/siginfo.h from pmon.h
libc6 renamed siginfo.h to siginfo-const.h sometime between
2.23 (in Xenial) and 2.27 (in bionic).

This builds on bionic and centos7 and in fact is required to
get DevStack to copmlete on bionic.

This is last in the stack since it has not been tested
beyond the compile/install that DevStack does.  There
may be a better/alternate solution...but with this we should
get a passing DevStack job.

Change-Id: I5a2ed9455b05e604731c3775d0f402c6137da2ef
Signed-off-by: Dean Troyer <dtroyer@gmail.com>