[WIP]fix zero sin6_scope_id issue
When running devstack on Ubuntu Bionic, if only ipv6 link-local address
is configured on interfaces, the calling of getaddrinfo with hint AF_INET6
in the function initAddr of msgClass.cpp succeed, but the sin6_scope_id
in the returned address structure is wrong. This will cause the failure of
the following bind operation. So we add this check, once the sin6_scope_id
is incorrect, make it switch to ipv4.
For normal StarlingX deployment, getaddrinfo will fail for the case of
ipv6, then will not run the check we added and switch to ipv4 directly.

Passed tests:
* Fresh building

Story: 2003161
Task: 29851

Change-Id: I25711acd70358042e0505051221091e32b0929e3
Signed-off-by: Yi Wang <yi.c.wang@intel.com>